Revert "Rename to component/"
This reverts commit c21dddcd026d235f3fdbb2613bafc3f538f25bd9.
