+++ lib/domain/record/components/notification_bar/components/ended_pill_sheet.dart
