+++ lib/domain/record/components/notification_bar/components/recommend_signup.dart
