+++ lib/domain/record/components/notification_bar/components/rest_duration.dart
