+++ test/domain/record/notification_bar/notification_bar_test.dart
@@ -1,15 +1,15 @@
 import 'package:pilll/analytics.dart';
 import 'package:pilll/domain/premium_introduction/util/discount_deadline.dart';
-import 'package:pilll/domain/record/components/notification_bar/component/discount_price_deadline.dart';
-import 'package:pilll/domain/record/components/notification_bar/component/ended_pill_sheet.dart';
-import 'package:pilll/domain/record/components/notification_bar/component/pilll_ads.dart';
+import 'package:pilll/domain/record/components/notification_bar/components/discount_price_deadline.dart';
+import 'package:pilll/domain/record/components/notification_bar/components/ended_pill_sheet.dart';
+import 'package:pilll/domain/record/components/notification_bar/components/pilll_ads.dart';
 import 'package:pilll/domain/record/components/notification_bar/notification_bar.dart';
 import 'package:pilll/domain/record/components/notification_bar/notification_bar_state.codegen.dart';
 import 'package:pilll/domain/record/components/notification_bar/state_notifier.dart';
-import 'package:pilll/domain/record/components/notification_bar/component/premium_trial_limit.dart';
-import 'package:pilll/domain/record/components/notification_bar/component/recommend_signup.dart';
-import 'package:pilll/domain/record/components/notification_bar/component/recommend_signup_premium.dart';
-import 'package:pilll/domain/record/components/notification_bar/component/rest_duration.dart';
+import 'package:pilll/domain/record/components/notification_bar/components/premium_trial_limit.dart';
+import 'package:pilll/domain/record/components/notification_bar/components/recommend_signup.dart';
+import 'package:pilll/domain/record/components/notification_bar/components/recommend_signup_premium.dart';
+import 'package:pilll/domain/record/components/notification_bar/components/rest_duration.dart';
 import 'package:pilll/entity/pill_sheet.codegen.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/entity/pill_sheet_type.dart';
